Show `Send as SMS` only if using Medic platform
Previously the `Send as SMS` button would show even if using a different
platform, but the Gateway number was only seen in the Medic platform
settings. This meant that if you were using ODK Aggregate you could see
SMS button if the gateway was previously set, but would not be able to
edit it. Changed the logic of showing button to match accordingly